+++ calculate.cpp
@@ -6,9 +6,9 @@
 string calculateBullAndPgia(string choice, string guess)
 {
     std::string ret = "";
-    size_t bull = 0;
+    uint bull = 0;
 
-    for (size_t i = 0; i < choice.size(); i++)
+    for (uint i = 0; i < choice.size(); i++)
     {
         if (choice[i] == guess[i])
             bull++;
